Split windows platform export template into multiple files
