+++ platform/windows/export/export.cpp
@@ -28,316 +28,11 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#include "core/io/file_access.h"
-#include "core/os/os.h"
-#include "editor/editor_export.h"
-#include "editor/editor_node.h"
-#include "editor/editor_settings.h"
-#include "platform/windows/logo.gen.h"
+#include "export.h"
+
+#include "export_plugin.h"
 
 static Error fixup_embedded_pck(const String &p_path, int64_t p_embedded_start, int64_t p_embedded_size);
-
-class EditorExportPlatformWindows : public EditorExportPlatformPC {
-	void _rcedit_add_data(const Ref<EditorExportPreset> &p_preset, const String &p_path);
-	Error _code_sign(const Ref<EditorExportPreset> &p_preset, const String &p_path);
-
-public:
-	virtual Error export_project(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path, int p_flags = 0);
-	virtual Error sign_shared_object(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path);
-	virtual void get_export_options(List<ExportOption> *r_options);
-};
-
-Error EditorExportPlatformWindows::sign_shared_object(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path) {
-	if (p_preset->get("codesign/enable")) {
-		return _code_sign(p_preset, p_path);
-	} else {
-		return OK;
-	}
-}
-
-Error EditorExportPlatformWindows::export_project(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path, int p_flags) {
-	Error err = EditorExportPlatformPC::export_project(p_preset, p_debug, p_path, p_flags);
-
-	if (err != OK) {
-		return err;
-	}
-
-	_rcedit_add_data(p_preset, p_path);
-
-	if (p_preset->get("codesign/enable") && err == OK) {
-		err = _code_sign(p_preset, p_path);
-	}
-
-	return err;
-}
-
-void EditorExportPlatformWindows::get_export_options(List<ExportOption> *r_options) {
-	EditorExportPlatformPC::get_export_options(r_options);
-
-	r_options->push_back(ExportOption(PropertyInfo(Variant::BOOL, "codesign/enable"), false));
-#ifdef WINDOWS_ENABLED
-	r_options->push_back(ExportOption(PropertyInfo(Variant::INT, "codesign/identity_type", PROPERTY_HINT_ENUM, "Select automatically,Use PKCS12 file (specify *.PFX/*.P12 file),Use certificate store (specify SHA1 hash)"), 0));
-#endif
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "codesign/identity", PROPERTY_HINT_GLOBAL_FILE, "*.pfx,*.p12"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "codesign/password"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::BOOL, "codesign/timestamp"), true));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "codesign/timestamp_server_url"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::INT, "codesign/digest_algorithm", PROPERTY_HINT_ENUM, "SHA1,SHA256"), 1));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "codesign/description"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::PACKED_STRING_ARRAY, "codesign/custom_options"), PackedStringArray()));
-
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "application/icon", PROPERTY_HINT_FILE, "*.ico"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "application/file_version", PROPERTY_HINT_PLACEHOLDER_TEXT, "1.0.0"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "application/product_version", PROPERTY_HINT_PLACEHOLDER_TEXT, "1.0.0"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "application/company_name", PROPERTY_HINT_PLACEHOLDER_TEXT, "Company Name"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "application/product_name", PROPERTY_HINT_PLACEHOLDER_TEXT, "Game Name"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "application/file_description"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "application/copyright"), ""));
-	r_options->push_back(ExportOption(PropertyInfo(Variant::STRING, "application/trademarks"), ""));
-}
-
-void EditorExportPlatformWindows::_rcedit_add_data(const Ref<EditorExportPreset> &p_preset, const String &p_path) {
-	String rcedit_path = EditorSettings::get_singleton()->get("export/windows/rcedit");
-
-	if (rcedit_path == String()) {
-		return;
-	}
-
-	if (!FileAccess::exists(rcedit_path)) {
-		ERR_PRINT("Could not find rcedit executable at " + rcedit_path + ", no icon or app information data will be included.");
-		return;
-	}
-
-#ifndef WINDOWS_ENABLED
-	// On non-Windows we need WINE to run rcedit
-	String wine_path = EditorSettings::get_singleton()->get("export/windows/wine");
-
-	if (wine_path != String() && !FileAccess::exists(wine_path)) {
-		ERR_PRINT("Could not find wine executable at " + wine_path + ", no icon or app information data will be included.");
-		return;
-	}
-
-	if (wine_path == String()) {
-		wine_path = "wine"; // try to run wine from PATH
-	}
-#endif
-
-	String icon_path = ProjectSettings::get_singleton()->globalize_path(p_preset->get("application/icon"));
-	String file_verion = p_preset->get("application/file_version");
-	String product_version = p_preset->get("application/product_version");
-	String company_name = p_preset->get("application/company_name");
-	String product_name = p_preset->get("application/product_name");
-	String file_description = p_preset->get("application/file_description");
-	String copyright = p_preset->get("application/copyright");
-	String trademarks = p_preset->get("application/trademarks");
-	String comments = p_preset->get("application/comments");
-
-	List<String> args;
-	args.push_back(p_path);
-	if (icon_path != String()) {
-		args.push_back("--set-icon");
-		args.push_back(icon_path);
-	}
-	if (file_verion != String()) {
-		args.push_back("--set-file-version");
-		args.push_back(file_verion);
-	}
-	if (product_version != String()) {
-		args.push_back("--set-product-version");
-		args.push_back(product_version);
-	}
-	if (company_name != String()) {
-		args.push_back("--set-version-string");
-		args.push_back("CompanyName");
-		args.push_back(company_name);
-	}
-	if (product_name != String()) {
-		args.push_back("--set-version-string");
-		args.push_back("ProductName");
-		args.push_back(product_name);
-	}
-	if (file_description != String()) {
-		args.push_back("--set-version-string");
-		args.push_back("FileDescription");
-		args.push_back(file_description);
-	}
-	if (copyright != String()) {
-		args.push_back("--set-version-string");
-		args.push_back("LegalCopyright");
-		args.push_back(copyright);
-	}
-	if (trademarks != String()) {
-		args.push_back("--set-version-string");
-		args.push_back("LegalTrademarks");
-		args.push_back(trademarks);
-	}
-
-#ifdef WINDOWS_ENABLED
-	OS::get_singleton()->execute(rcedit_path, args);
-#else
-	// On non-Windows we need WINE to run rcedit
-	args.push_front(rcedit_path);
-	OS::get_singleton()->execute(wine_path, args);
-#endif
-}
-
-Error EditorExportPlatformWindows::_code_sign(const Ref<EditorExportPreset> &p_preset, const String &p_path) {
-	List<String> args;
-
-#ifdef WINDOWS_ENABLED
-	String signtool_path = EditorSettings::get_singleton()->get("export/windows/signtool");
-	if (signtool_path != String() && !FileAccess::exists(signtool_path)) {
-		ERR_PRINT("Could not find signtool executable at " + signtool_path + ", aborting.");
-		return ERR_FILE_NOT_FOUND;
-	}
-	if (signtool_path == String()) {
-		signtool_path = "signtool"; // try to run signtool from PATH
-	}
-#else
-	String signtool_path = EditorSettings::get_singleton()->get("export/windows/osslsigncode");
-	if (signtool_path != String() && !FileAccess::exists(signtool_path)) {
-		ERR_PRINT("Could not find osslsigncode executable at " + signtool_path + ", aborting.");
-		return ERR_FILE_NOT_FOUND;
-	}
-	if (signtool_path == String()) {
-		signtool_path = "osslsigncode"; // try to run signtool from PATH
-	}
-#endif
-
-	args.push_back("sign");
-
-	//identity
-#ifdef WINDOWS_ENABLED
-	int id_type = p_preset->get("codesign/identity_type");
-	if (id_type == 0) { //auto select
-		args.push_back("/a");
-	} else if (id_type == 1) { //pkcs12
-		if (p_preset->get("codesign/identity") != "") {
-			args.push_back("/f");
-			args.push_back(p_preset->get("codesign/identity"));
-		} else {
-			EditorNode::add_io_error("codesign: no identity found");
-			return FAILED;
-		}
-	} else if (id_type == 2) { //Windows certificate store
-		if (p_preset->get("codesign/identity") != "") {
-			args.push_back("/sha1");
-			args.push_back(p_preset->get("codesign/identity"));
-		} else {
-			EditorNode::add_io_error("codesign: no identity found");
-			return FAILED;
-		}
-	} else {
-		EditorNode::add_io_error("codesign: invalid identity type");
-		return FAILED;
-	}
-#else
-	if (p_preset->get("codesign/identity") != "") {
-		args.push_back("-pkcs12");
-		args.push_back(p_preset->get("codesign/identity"));
-	} else {
-		EditorNode::add_io_error("codesign: no identity found");
-		return FAILED;
-	}
-#endif
-
-	//password
-	if (p_preset->get("codesign/password") != "") {
-#ifdef WINDOWS_ENABLED
-		args.push_back("/p");
-#else
-		args.push_back("-pass");
-#endif
-		args.push_back(p_preset->get("codesign/password"));
-	}
-
-	//timestamp
-	if (p_preset->get("codesign/timestamp")) {
-		if (p_preset->get("codesign/timestamp_server") != "") {
-#ifdef WINDOWS_ENABLED
-			args.push_back("/tr");
-			args.push_back(p_preset->get("codesign/timestamp_server_url"));
-			args.push_back("/td");
-			if ((int)p_preset->get("codesign/digest_algorithm") == 0) {
-				args.push_back("sha1");
-			} else {
-				args.push_back("sha256");
-			}
-#else
-			args.push_back("-ts");
-			args.push_back(p_preset->get("codesign/timestamp_server_url"));
-#endif
-		} else {
-			EditorNode::add_io_error("codesign: invalid timestamp server");
-			return FAILED;
-		}
-	}
-
-	//digest
-#ifdef WINDOWS_ENABLED
-	args.push_back("/fd");
-#else
-	args.push_back("-h");
-#endif
-	if ((int)p_preset->get("codesign/digest_algorithm") == 0) {
-		args.push_back("sha1");
-	} else {
-		args.push_back("sha256");
-	}
-
-	//description
-	if (p_preset->get("codesign/description") != "") {
-#ifdef WINDOWS_ENABLED
-		args.push_back("/d");
-#else
-		args.push_back("-n");
-#endif
-		args.push_back(p_preset->get("codesign/description"));
-	}
-
-	//user options
-	PackedStringArray user_args = p_preset->get("codesign/custom_options");
-	for (int i = 0; i < user_args.size(); i++) {
-		String user_arg = user_args[i].strip_edges();
-		if (!user_arg.is_empty()) {
-			args.push_back(user_arg);
-		}
-	}
-
-#ifndef WINDOWS_ENABLED
-	args.push_back("-in");
-#endif
-	args.push_back(p_path);
-#ifndef WINDOWS_ENABLED
-	args.push_back("-out");
-	args.push_back(p_path + "_signed");
-#endif
-
-	String str;
-	Error err = OS::get_singleton()->execute(signtool_path, args, &str, nullptr, true);
-	ERR_FAIL_COND_V(err != OK, err);
-
-	print_line("codesign (" + p_path + "): " + str);
-#ifndef WINDOWS_ENABLED
-	if (str.find("SignTool Error") != -1) {
-#else
-	if (str.find("Failed") != -1) {
-#endif
-		return FAILED;
-	}
-
-#ifndef WINDOWS_ENABLED
-	DirAccessRef tmp_dir = DirAccess::create_for_path(p_path.get_base_dir());
-
-	err = tmp_dir->remove(p_path);
-	ERR_FAIL_COND_V(err != OK, err);
-
-	err = tmp_dir->rename(p_path + "_signed", p_path);
-	ERR_FAIL_COND_V(err != OK, err);
-#endif
-
-	return OK;
-}
 
 void register_windows_exporter() {
 	EDITOR_DEF("export/windows/rcedit", "");
